Made the cookie clicker leaderboard number more exact

+++ html/api/cookieClicker.php
@@ -1,18 +1,19 @@
 <?php
+$skipThrottle = true;
 require_once "api.php";
 header("Access-Control-Allow-Origin: *");
-if ($_POST["type"] === "view" and $_POST["username"] !== null and $_POST["cookies"] !== null and $_POST["cookiesPs"] !== null and $_POST["room"] !== null) {
+if ($_POST["type"] === "view" and $_POST["username"] !== null and $_POST["cookies"] !== null and $_POST["cookiesPs"] !== null and $_POST["room"] !== null and $_POST["time"] !== null) {
     $username = $_POST["username"];
     $cookies = intval($_POST["cookies"]);
     $cookiesPs = intval($_POST["cookiesPs"]);
     $room = $_POST["room"];
-    $time = time();
+    $time = intval($_POST["time"]);
     if (dbRequest2("SELECT * FROM cookieClicker WHERE room='$room' AND username='$username'")) {
-        dbCommand("UPDATE cookieClicker SET cookies = '$cookies', cookiesPerSecond = '$cookiesPs', lastUpdate='$time' WHERE room='$room' AND username='$username';");
+        dbCommand("UPDATE cookieClicker SET cookies = '$cookies', cookiesPs = '$cookiesPs', lastUpdate='$time' WHERE room='$room' AND username='$username';");
     } else {
-        dbCommand("INSERT INTO cookieClicker (username, room, cookies, cookiesPerSecond, lastUpdate) VALUES ('$username', '$room', $cookies, $cookiesPs, $time)");
+        dbCommand("INSERT INTO cookieClicker (username, room, cookies, cookiesPs, lastUpdate) VALUES ('$username', '$room', $cookies, $cookiesPs, $time)");
     }
-    echo json_encode(["leaderboard" => dbRequest2("SELECT * FROM cookieClicker WHERE room='$room' ORDER BY cookiesPerSecond DESC, cookies DESC"), "commands" => dbRequest2("SELECT * FROM cookieClickerCommand WHERE room='$room' AND username='$username'")]);
+    echo json_encode(["leaderboard" => dbRequest2("SELECT * FROM cookieClicker WHERE room='$room' ORDER BY cookiesPs DESC, cookies DESC"), "commands" => dbRequest2("SELECT * FROM cookieClickerCommand WHERE room='$room' AND username='$username'")]);
     dbCommand("DELETE FROM cookieClickerCommand WHERE room='$room' AND username='$username'");
 } elseif ($_POST["type"] === "donate" and $_POST["username"] !== null and $_POST["room"] !== null and $_POST["cookies"] !== null) {
     $username = $_POST["username"];
